fix: Admin index base

+++ database/migrations/2021_01_01_144137_create_sessions_table.php
@@ -11,7 +11,7 @@ class CreateSessionsTable extends Migration
         Schema::create('sessions', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->foreignId('session_stream_id');
-            $table->integer('zoom_meeting_id');
+            $table->bigInteger('zoom_meeting_id');
             $table->string('title');
             $table->longText('description');
             $table->dateTimeTz('start');
